Rolled back to May 2, 2018 version
Two problems exist with the homepage: 1) the video loops after a couple of seconds, and 2) the navigation arrows for changing country (+ video + flag + welcome headline) don't work. The client wants a quick fix so we simply rolled back to a version that works without checking and fixing any other parts of the code.